# Conflicts:
#	app/src/main/java/com/petzoo/petzoo/MapsFragment.java
#	app/src/main/java/com/petzoo/petzoo/models/Alerta.java

+++ app/src/main/java/com/petzoo/petzoo/models/Alerta.java
@@ -1,15 +1,15 @@
 package com.petzoo.petzoo.models;
 
 public class Alerta {
-    private float IdAlerta;
+    private int IdAlerta;
     private String Descripcion;
     private float Latitud;
     private float Longitud;
     private String FechaAlerta;
-    private float IdPersona;
-    private float IdPersonaRescate;
+    private int IdPersona;
+    private int IdPersonaRescate;
     private int IdClaseMascota;
-    private float IdEstadoAlerta;
+    private int IdEstadoAlerta;
     //ArrayList < Object > Foto = new ArrayList < Object > ();
     //private String ClaseMascota = null;
     //private String EstadoAlerta = null;
@@ -19,7 +19,7 @@ public class Alerta {
 
     // Getter Methods
 
-    public float getIdAlerta() {
+    public int getIdAlerta() {
         return IdAlerta;
     }
 
@@ -39,11 +39,11 @@ public class Alerta {
         return FechaAlerta;
     }
 
-    public float getIdPersona() {
+    public int getIdPersona() {
         return IdPersona;
     }
 
-    public float getIdPersonaRescate() {
+    public int getIdPersonaRescate() {
         return IdPersonaRescate;
     }
 
@@ -51,7 +51,7 @@ public class Alerta {
         return IdClaseMascota;
     }
 
-    public float getIdEstadoAlerta() {
+    public int getIdEstadoAlerta() {
         return IdEstadoAlerta;
     }
 /*
@@ -73,7 +73,7 @@ public class Alerta {
 */
     // Setter Methods
 
-    public void setIdAlerta(float IdAlerta) {
+    public void setIdAlerta(int IdAlerta) {
         this.IdAlerta = IdAlerta;
     }
 
@@ -93,11 +93,11 @@ public class Alerta {
         this.FechaAlerta = FechaAlerta;
     }
 
-    public void setIdPersona(float IdPersona) {
+    public void setIdPersona(int IdPersona) {
         this.IdPersona = IdPersona;
     }
 
-    public void setIdPersonaRescate(float IdPersonaRescate) {
+    public void setIdPersonaRescate(int IdPersonaRescate) {
         this.IdPersonaRescate = IdPersonaRescate;
     }
 
@@ -105,7 +105,7 @@ public class Alerta {
         this.IdClaseMascota = IdClaseMascota;
     }
 
-    public void setIdEstadoAlerta(float IdEstadoAlerta) {
+    public void setIdEstadoAlerta(int IdEstadoAlerta) {
         this.IdEstadoAlerta = IdEstadoAlerta;
     }
 /*
